requirements.txt
paho-mqtt>=2.0.0
Pillow>=10.0.0
Flask>=3.0.0
RPi.GPIO
spidev
waveshare-epd @ git+https://github.com/waveshare/e-Paper.git#subdirectory=RaspberryPi_JetsonNano/python
